+++ spec/playground_spec.rb
@@ -4,36 +4,6 @@ require "shiny_json_logic/errors/base"
 RSpec.describe ShinyJsonLogic do
   describe "standard behavior" do
     [
-      {
-        "description" => "Filter, map, all, none, and some (8)",
-        "rule" => {
-          "reduce" => [
-            {
-              "var" => "integers"
-            },
-            {
-              "+" => [
-                {
-                  "var" => "current"
-                },
-                {
-                  "var" => "accumulator"
-                }
-              ]
-            },
-            0
-          ]
-        },
-        "data" => {
-          "integers" => [
-            1,
-            2,
-            3,
-            4
-          ]
-        },
-        "result" => 10
-      }
     ].each_with_index do |testcase, index|
       next unless testcase.is_a?(Hash)